Move mock classes to static inner
